ci: add configure-docker action
Another action to reduce boilerplate in the main ci.yml file.

This action will set up a docker builder compatible with caching build
layers to a container registry using the `docker-container` build
driver.

The action will then determine whether it should log in to a container
registry or not based on whether a username and token have been
provided. This will occur only on `push` to master branch, as the
`pull_request` GitHub context does not allow the use of repository
secrets.

Finally it will configure the docker build cache args.

When logged in build cache layers will be both pulled and pushed to the
registry.

When not logged in build cached layers will only be pulled from the
registry.